replace old pdfreader with newreader
