--- src/NewReader/PdfReader.css
@@ -1,369 +1,0 @@
-.pdf-root {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-  min-height: 400px;
-  background: #525659;
-  color: #e8eaed;
-  overflow: hidden;
-}
-
-.pdf-root:fullscreen {
-  width: 100vw;
-  height: 100vh;
-}
-
-.pdf-toolbar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 8px;
-  padding: 6px 10px;
-  background: #2b2d30;
-  border-bottom: 1px solid #1c1d1f;
-  flex-wrap: wrap;
-  flex-shrink: 0;
-}
-
-.pdf-toolbar-group {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-}
-
-.pdf-btn {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  border: none;
-  border-radius: 6px;
-  background: transparent;
-  color: #e8eaed;
-  cursor: pointer;
-  transition: background 0.12s ease;
-}
-
-.pdf-btn:hover:not(:disabled) {
-  background: rgba(255, 255, 255, 0.1);
-}
-
-.pdf-btn:disabled {
-  opacity: 0.35;
-  cursor: default;
-}
-
-.pdf-page-form {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  margin: 0 2px;
-}
-
-.pdf-page-input {
-  width: 40px;
-  text-align: center;
-  background: #1c1d1f;
-  border: 1px solid #3c4043;
-  border-radius: 4px;
-  color: #e8eaed;
-  font-size: 13px;
-  padding: 4px 2px;
-}
-
-.pdf-page-input:focus {
-  outline: 1px solid #8ab4f8;
-}
-
-.pdf-page-total,
-.pdf-zoom-label {
-  font-size: 13px;
-  color: #b0b3b8;
-  min-width: 42px;
-  text-align: center;
-  user-select: none;
-}
-
-/* Body: sidebar + viewport */
-.pdf-body {
-  display: flex;
-  flex: 1;
-  min-height: 0;
-}
-
-.pdf-sidebar {
-  width: 260px;
-  flex-shrink: 0;
-  background: #323639;
-  border-right: 1px solid #1c1d1f;
-  overflow-y: auto;
-}
-
-.pdf-sidebar-header {
-  padding: 12px 14px;
-  font-size: 12px;
-  font-weight: 600;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-  color: #9aa0a6;
-  border-bottom: 1px solid #1c1d1f;
-}
-
-.pdf-sidebar-empty {
-  padding: 14px;
-  font-size: 13px;
-  color: #7a7d81;
-}
-
-.pdf-toc-list {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.pdf-toc-item {
-  display: block;
-  width: 100%;
-  text-align: left;
-  background: none;
-  border: none;
-  color: #d3d6db;
-  font-size: 13px;
-  padding: 7px 12px;
-  cursor: pointer;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.pdf-toc-item:hover:not(:disabled) {
-  background: rgba(255, 255, 255, 0.06);
-}
-
-.pdf-toc-item.active {
-  background: rgba(66, 133, 244, 0.25);
-  color: #8ab4f8;
-}
-
-.pdf-toc-item:disabled {
-  color: #6a6d71;
-  cursor: default;
-}
-
-/* Viewport (scroll container) */
-.pdf-viewport {
-  flex: 1;
-  overflow: auto;
-  position: relative;
-  background: #252525;
-  color-scheme: dark;
-}
-
-/* Loading / error states */
-.pdf-status-container {
-  position: absolute;
-  inset: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.pdf-status {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
-  color: #b0b3b8;
-  font-size: 14px;
-}
-
-.pdf-error {
-  color: #f28b82;
-}
-
-.pdf-pages-stack {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  width: max-content;
-  min-width: 100%;
-  padding-top: 16px;
-  background: #252525;
-}
-
-.pdf-page-wrap {
-  position: relative;
-  flex-shrink: 0;
-  margin-inline: auto;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.5);
-  background: #252525;
-}
-
-.pdf-canvas {
-  display: block;
-}
-
-/* Text and annotation layer styles adapted from react-pdf/pdf.js */
-.pdf-root {
-  --react-pdf-text-layer: 1;
-}
-
-[data-main-rotation='90'] .pdf-text-layer.textLayer {
-  transform: rotate(90deg) translateY(-100%);
-}
-
-[data-main-rotation='180'] .pdf-text-layer.textLayer {
-  transform: rotate(180deg) translate(-100%, -100%);
-}
-
-[data-main-rotation='270'] .pdf-text-layer.textLayer {
-  transform: rotate(270deg) translateX(-100%);
-}
-
-.pdf-text-layer.textLayer {
-  position: absolute;
-  text-align: initial;
-  inset: 0;
-  overflow: clip;
-  opacity: 1;
-  line-height: 1;
-  text-size-adjust: none;
-  forced-color-adjust: none;
-  transform-origin: 0 0;
-  caret-color: CanvasText;
-  z-index: 0;
-
-  --min-font-size: 1;
-  --total-scale-factor: var(--scale-factor, 1);
-  --text-scale-factor: calc(var(--total-scale-factor) * var(--min-font-size));
-  --min-font-size-inv: calc(1 / var(--min-font-size));
-}
-
-.pdf-text-layer.textLayer :is(span, br) {
-  color: transparent;
-  position: absolute;
-  white-space: pre;
-  cursor: text;
-  margin: 0;
-  transform-origin: 0 0;
-}
-
-.pdf-text-layer.textLayer > :not(.markedContent),
-.pdf-text-layer.textLayer .markedContent span:not(.markedContent) {
-  z-index: 1;
-
-  --font-height: 0;
-  font-size: calc(var(--text-scale-factor) * var(--font-height));
-
-  --scale-x: 1;
-  --rotate: 0deg;
-  transform: rotate(var(--rotate)) scaleX(var(--scale-x))
-    scale(var(--min-font-size-inv));
-}
-
-.pdf-text-layer.textLayer .markedContent {
-  display: contents;
-}
-
-.pdf-text-layer.textLayer span[role='img'] {
-  user-select: none;
-  cursor: default;
-}
-
-.pdf-text-layer.textLayer .highlight {
-  margin: -1px;
-  padding: 1px;
-  border-radius: 4px;
-}
-
-.pdf-text-layer.textLayer .highlight.appended {
-  position: initial;
-}
-
-.pdf-text-layer.textLayer .highlight.begin {
-  border-radius: 4px 0 0 4px;
-}
-
-.pdf-text-layer.textLayer .highlight.end {
-  border-radius: 0 4px 4px 0;
-}
-
-.pdf-text-layer.textLayer .highlight.middle {
-  border-radius: 0;
-}
-
-.pdf-text-layer.textLayer br::selection {
-  background: transparent;
-}
-
-.pdf-text-layer.textLayer .endOfContent {
-  display: block;
-  position: absolute;
-  inset: 100% 0 0;
-  z-index: 0;
-  cursor: default;
-  user-select: none;
-}
-
-.pdf-text-layer.textLayer.selecting .endOfContent {
-  top: 0;
-}
-
-.pdf-text-layer.textLayer .hiddenCanvasElement {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 0;
-  height: 0;
-  display: none;
-}
-
-.pdf-annotation-layer {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  overflow: hidden;
-}
-
-.pdf-annotation-layer section {
-  position: absolute;
-}
-
-.pdf-annotation-layer .linkAnnotation > a {
-  position: absolute;
-  font-size: 1em;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-}
-
-.pdf-annotation-layer .linkAnnotation > a:hover {
-  background: rgba(255, 255, 0, 0.15);
-  box-shadow: 0 0 0 1px rgba(255, 255, 0, 0.4);
-}
-
-.pdf-annotation-layer input,
-.pdf-annotation-layer textarea,
-.pdf-annotation-layer select {
-  background: rgba(255, 255, 255, 0.6);
-  border: 1px solid #8ab4f8;
-  box-sizing: border-box;
-  font: inherit;
-}
-
-@media (max-width: 640px) {
-  .pdf-sidebar {
-    position: absolute;
-    z-index: 5;
-    height: 100%;
-  }
-  .pdf-page-total,
-  .pdf-zoom-label {
-    min-width: unset;
-  }
-}
