+++ src/PdfReader/types.ts
@@ -1,68 +1,63 @@
-import {
-  ActiveReaderArguments,
-  FitMode,
-  InactiveReaderArguments,
-  ReaderSettings,
-  ReaderState,
-} from '../types';
+import { Dispatch } from 'react';
+import { WebpubManifest } from '../types';
+import { PdfReaderAction } from './reducer';
 
-export type InternalState = {
-  resourceIndex: number;
-  resource: { data: Uint8Array } | null;
-  // we only know the numPages once the resource has been parsed
-  numPages: number | null;
-  // if pageNumber is -1, we will navigate to the end of the
-  // resource once it is parsed
+export type FitMode = 'width' | 'height' | null;
+
+export interface OutlineItem {
+  title: string;
+  pageNumber: number | null;
+  items: OutlineItem[];
+}
+
+export interface PageSize {
+  width: number;
+  height: number;
+}
+
+export interface ViewportAnchor {
   pageNumber: number;
+  intraPageRatio: number;
+  viewportOffset: number;
+}
+
+export interface PdfReaderProps {
+  fileUrl?: string;
+  webpubManifestUrl?: string;
+  manifest?: WebpubManifest;
+  proxyUrl?: string;
+  pdfWorkerSrc?: string;
+  height?: string | number;
+  className?: string;
+  toggleFullScreen?: () => void;
+}
+
+export interface PdfReaderContentProps {
+  fileUrl: string | undefined;
+  pdfWorkerSrc?: string;
+  pageNumber: number;
+  navigationRequestId: number;
   scale: number;
-  pdfHeight: number;
-  pdfWidth: number;
-  pageHeight: number | undefined;
-  pageWidth: number | undefined;
-  rendered: boolean;
-};
+  fitMode: FitMode;
+  rotation: number;
+  dispatch: Dispatch<PdfReaderAction>;
+  pendingAction: PdfReaderAction | null;
+  clearPendingAction: () => void;
+  onOutlineLoad: (outlineItems: OutlineItem[]) => void;
+  onPageSizesReady?: () => void;
+}
 
-export type InactiveState = ReaderState &
-  InternalState & { state: 'INACTIVE'; settings: undefined };
+export interface RenderTask {
+  promise: Promise<unknown>;
+  cancel: () => void;
+}
 
-export type ActiveState = ReaderState &
-  InternalState & { state: 'ACTIVE'; settings: ReaderSettings };
+export interface PdfOutlineEntry {
+  title: string;
+  dest: string | unknown[] | null;
+  items?: PdfOutlineEntry[];
+}
 
-export type ErrorState = ReaderState &
-  InternalState & {
-    state: 'ERROR';
-    error: Error;
-    settings: ReaderSettings;
-  };
+export type RefProxy = import('pdfjs-dist/types/src/display/api').RefProxy;
 
-export type PdfState = InactiveState | ActiveState | ErrorState;
-
-export type PdfReaderArguments =
-  | ActiveReaderArguments<Uint8Array>
-  | InactiveReaderArguments;
-
-export type PdfReaderAction =
-  | {
-      type: 'ARGS_CHANGED';
-      args: PdfReaderArguments;
-    }
-  | { type: 'GO_FORWARD' }
-  | { type: 'GO_BACKWARD' }
-  | { type: 'GO_TO_HREF'; href: string }
-  | { type: 'GO_TO_PAGE'; page: number }
-  | { type: 'PAGE_IN_VIEW'; page: number }
-  | { type: 'RESOURCE_FETCH_SUCCESS'; resource: { data: Uint8Array } }
-  | { type: 'PDF_PARSED'; numPages: number }
-  | { type: 'PDF_LOAD_ERROR'; error: Error }
-  | { type: 'SET_SCALE'; scale: number }
-  // | { type: 'RESET_SETTINGS' } may be needed in future
-  | { type: 'SET_SCROLL'; isScrolling: boolean }
-  | { type: 'PAGE_LOAD_SUCCESS'; height: number; width: number }
-  | {
-      type: 'RESIZE_PAGE';
-      height: number | undefined;
-      width: number | undefined;
-    }
-  | { type: 'SET_FIT_MODE'; fitMode: FitMode }
-  | { type: 'ROTATE_COUNTER_CLOCKWISE' }
-  | { type: 'BOOK_BOUNDARY_CHANGED'; atStart: boolean; atEnd: boolean };
+export type PdfBinaryInput = Blob | File | Uint8Array | ArrayBuffer;
